backend/requirements.txt
flask
flask-cors
python-docx
google-generativeai
python-dotenv
werkzeug
